refactor(cli): extract HQ telemetry bridges + dep-watcher consumers from cli-main.ts
